pyproject.toml
[tool.poetry]
name = "ProbaProjekat"
version = "0.1.0"
description = "Poetry proba"
authors = ["vuk"]

[tool.poetry.dependencies]
python = "3.7.10"
numpy = "1.18.*"
Django="^3.2.9"
djangorestframework = "^3.12.4"
markdown = "^3.*"
django-filter="^21.*"
psycopg2-binary= "^2.9.*"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
